Improved input validation. Improved documentation.
Prevented negative length from producing a huge string and possibly exhausting all of memory. When using custom mask chars property of options object is checked that is a string. Non custom pattern is checked that is a valid option from the list of possibilities. Removed unnecessary parseInt call when creating the result. Adding new options for string creation is now simpler and does not require changes to multiple places. Type checking of parameter pattern now allows only string or a number.